# openapi.yml
openapi: 3.0.0
info:
  description: Simple book service API
  title: Bookshelf API
  version: 0.0.1
paths:
  /book:
    get:
      description: Get all books
      responses:
        '200':
          description: Return all books in JSON format
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookDTO'
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookDTO'
      responses:
        '201':
          description: Return created book in Json format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDTO'
        '400':
          description: Invalid book object send in request body
  /book/{id}:
    get:
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric Id of the book to get
      responses:
        '200':
          description: Return book in Json format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDTO'
        '404':
          description: No user with given id found
    put:
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric Id of the book to get
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookDTO'
      responses:
        '200':
          description: Return updated book in Json format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDTO'
        '400':
          description: Invalid book object send in request body
    delete:
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric Id of the book to get
      responses:
        '200':
          description: Empty response body

components:
  schemas:
    BookDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        title:
          type: string
        author:
          type: string
        isbnnumber:
          type: string
          description: Valid ISBN number
        pages:
          type: integer
          format: int32
        rate:
          type: integer
          minimum: 0
          maximum: 5